move navitems
