+++ apps/web/src/components/Layout/MobileSlider.tsx
@@ -4,29 +4,19 @@ import { useTranslation } from 'react-i18next';
 
 import { Navigation } from './Navigation';
 
-import type { NavItem } from './types';
-
 export type MobileSliderProps = {
-  activeItemId?: string;
   isOpen: boolean;
-  items: NavItem[] | NavItem[][];
   onNavigate?: (id: string) => void;
   setIsOpen: (isOpen: boolean) => void;
 };
 
-export const MobileSlider = ({ activeItemId, isOpen, items, onNavigate, setIsOpen }: MobileSliderProps) => {
+export const MobileSlider = ({ isOpen, onNavigate, setIsOpen }: MobileSliderProps) => {
   const { i18n } = useTranslation();
   return (
     <Slider isOpen={isOpen} setIsOpen={setIsOpen} title={<Branding />}>
       <div className="flex h-full flex-col">
         <div className="flex-grow">
-          <Navigation
-            activeItemId={activeItemId}
-            isAlwaysDark={false}
-            items={items}
-            orientation="vertical"
-            onNavigate={onNavigate}
-          />
+          <Navigation isAlwaysDark={false} orientation="vertical" onNavigate={onNavigate} />
         </div>
         <div className="flex items-center justify-between text-slate-700 dark:text-slate-300">
           <button
