--- apps/web/src/components/Layout/Navbar.stories.tsx
@@ -1,27 +1,0 @@
-import { range } from '@douglasneuroinformatics/utils';
-import { QuestionMarkCircleIcon } from '@heroicons/react/24/outline';
-import type { Meta, StoryObj } from '@storybook/react';
-
-import { Navbar } from './Navbar';
-
-type Story = StoryObj<typeof Navbar>;
-
-export default {
-  component: Navbar,
-  parameters: {
-    layout: 'fullscreen'
-  }
-} satisfies Meta<typeof Navbar>;
-
-export const Default: Story = {
-  args: {
-    activeItemId: '1',
-    items: range(1, 5).map((i) => ({
-      icon: QuestionMarkCircleIcon,
-      id: i.toString(),
-      label: `Page ${i}`,
-      onClick: () => undefined
-    })),
-    onNavigate: (id) => alert(`Navigate to ID: ${id}`)
-  }
-};
